Fix develop build failure by creating a common test base for transfor… (#286)
* Fix develop build failure by creating a common test base for transform module and add all the plugins used to the plugin artifact

* Ignore copybook tests since CDAP-6162 is blocking it
