+++ copybookreader-plugins/src/test/java/co/cask/hydrator/plugin/CopybookTest.java
@@ -57,6 +57,7 @@ import org.junit.AfterClass;
 import org.junit.Assert;
 import org.junit.BeforeClass;
 import org.junit.ClassRule;
+import org.junit.Ignore;
 import org.junit.Test;
 import org.junit.rules.TemporaryFolder;
 
@@ -71,6 +72,7 @@ import java.util.concurrent.TimeUnit;
 /**
  * Unit test for {@link CopybookSource} classes.
  */
+@Ignore
 public class CopybookTest extends HydratorTestBase {
 
   @ClassRule
